Make the ThemeService static.
Implement the ThemeService the same way Xamarin.Essentials implements their own cross platform extensions.
