+++ src/OpenExtensions/Core/Services/ThemeService/ThemeService.netstandard.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Threading.Tasks;
+
+namespace OpenExtensions.Core.Services
+{
+    public static partial class ThemeService
+    {
+        static ThemeType GetPlatformCurrentTheme() => throw new NotImplementedException();
+        static Task SetPlatformTheme(ThemeType theme) => throw new NotImplementedException();
+    }
+}
