+++ src/OpenExtensions/Core/Services/ThemeService/ThemeType.shared.cs
@@ -1,0 +1,21 @@
+﻿namespace OpenExtensions.Core.Services
+{
+    /// <summary>
+    /// The type of the theme used in <see cref="ThemeService"/>
+    /// Enum integer values are the same as UWP ElementTheme values.
+    /// </summary>
+    public enum ThemeType
+    {
+        /// <summary>
+        /// </summary>
+        Default = 0,
+
+        /// <summary>
+        /// </summary>
+        Light = 1,
+
+        /// <summary>
+        /// </summary>
+        Dark = 2
+    }
+}
